refactor(DWS): move the API which query the dws flavor to v2/cluster
