+++ openstack/dws/v1/cluster/results.go
@@ -106,22 +106,6 @@ type ResizeInfo struct {
 	OriginNodeNum int    `json:"origin_node_num"`
 }
 
-type NodeTypes struct {
-	NodeTypes []NodeType `json:"node_types"`
-}
-
-type NodeType struct {
-	Detail   []TypeDetail `json:"detail"`
-	Id       string       `json:"id"`
-	SpecName string       `json:"spec_name"`
-}
-
-type TypeDetail struct {
-	Unit  string `json:"unit"`
-	Type  string `json:"type"`
-	Value string `json:"value"`
-}
-
 type GetResult struct {
 	golangsdk.Result
 }
