+++ openstack/dws/v2/flavors/urls.go
@@ -1,0 +1,8 @@
+package flavors
+
+import "github.com/chnsz/golangsdk"
+
+// listNodeTypesURL /v2/{project_id}/node-types
+func listNodeTypesURL(c *golangsdk.ServiceClient) string {
+	return c.ServiceURL("node-types")
+}
